ml-ops-2/fraud_detector/requirements.txt
kafka-python
pandas
numpy
catboost